dashboard_analyzer/data_collection/test_token_refresh_simulation.py
#!/usr/bin/env python3
"""
Token Refresh Simulation Test
Simulates the complete token refresh process without requiring Chrome
"""

import sys
import os
import time
import jwt
from datetime import datetime, timedelta, timezone

# Add paths for imports
sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

from token_manager import TokenManager
from chatbot_verbatims_collector import ChatbotVerbatimsCollector


def generate_fresh_token():
    """Generate a fresh JWT token for testing"""
    
    # Create payload similar to your real token
    now = datetime.now(timezone.utc)
    payload = {
        "aud": "aeae33a8-e44a-4287-a4d8-005df76ee97d",
        "iss": "https://login.microsoftonline.com/188b450b-7545-493f-9096-854ed4977730/v2.0",
        "iat": int(now.timestamp()),
        "nbf": int(now.timestamp()),
        "exp": int((now + timedelta(hours=1)).timestamp()),  # Valid for 1 hour
        "email": "KMM.drivera@iberia.es",
        "family_name": "Rivera López-Brea",
        "given_name": "Diego",
        "name": "KMM.Rivera López-Brea, Diego",
        "oid": "d6a8284d-f35b-42ef-bf0a-1acc30e75e28",
        "preferred_username": "KMM.drivera@iberia.es",
        "scp": "user_impersonation",
        "tid": "188b450b-7545-493f-9096-854ed4977730",
        "ver": "2.0"
    }
    
    # Generate a mock JWT (not cryptographically signed, just for testing)
    header = "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImtpZCI6IkpZaEFjVFBNWl9MWDZEQmxPV1E3SG4wTmVYRSJ9"
    
    import json
    import base64
    
    # Encode payload
    payload_json = json.dumps(payload, separators=(',', ':'))
    payload_b64 = base64.urlsafe_b64encode(payload_json.encode()).decode().rstrip('=')
    
    # Create signature (mock)
    signature = "MOCK_SIGNATURE_FOR_TESTING_PURPOSES_ONLY_123456789"
    
    return f"{header}.{payload_b64}.{signature}"


def simulate_token_refresh():
    """Simulate the complete token refresh process"""
    
    print("🚀 Token Refresh Simulation Test")
    print("=" * 60)
    
    # Step 1: Initialize TokenManager
    print("\n📋 Step 1: Initialize TokenManager")
    token_manager = TokenManager()
    
    # Step 2: Check current token status
    print("\n📋 Step 2: Check current token status")
    current_info = token_manager.get_token_info()
    print(f"📊 Current status: {current_info['status']}")
    print(f"⏰ Expires in: {current_info.get('expires_in', 'N/A')} seconds")
    
    # Step 3: Verify token needs refresh
    print("\n📋 Step 3: Verify token needs refresh")
    needs_refresh = current_info.get('needs_refresh', True)
    print(f"🔄 Needs refresh: {needs_refresh}")
    
    if not needs_refresh and current_info['status'] == 'valid':
        print("✅ Token is still valid, forcing expiration for test...")
        # Simulate an expired token by modifying the file
        old_token = token_manager._load_token_from_file()
        if old_token:
            # Decode and modify expiration
            try:
                decoded = jwt.decode(old_token, options={"verify_signature": False})
                # Set expiration to 1 minute ago
                decoded['exp'] = int((datetime.now(timezone.utc) - timedelta(minutes=1)).timestamp())
                
                # Create new expired token for testing
                import json
                import base64
                header = "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImtpZCI6IkpZaEFjVFBNWl9MWDZEQmxPV1E3SG4wTmVYRSJ9"
                payload_json = json.dumps(decoded, separators=(',', ':'))
                payload_b64 = base64.urlsafe_b64encode(payload_json.encode()).decode().rstrip('=')
                signature = "EXPIRED_TOKEN_FOR_TESTING"
                expired_token = f"{header}.{payload_b64}.{signature}"
                
                # Save expired token
                token_manager._save_token_to_file(expired_token)
                print("⚠️ Token artificially expired for testing")
                
            except Exception as e:
                print(f"Warning: Could not modify token: {e}")
    
    # Step 4: Simulate Chrome automation finding new token
    print("\n📋 Step 4: Simulate Chrome automation extracting fresh token")
    print("🌐 [SIMULATION] Opening Chrome browser...")
    time.sleep(1)
    print("🧭 [SIMULATION] Navigating to https://nps.chatbot.iberia.es/home...")
    time.sleep(1)
    print("📡 [SIMULATION] Waiting for network requests...")
    time.sleep(1)
    print("🔍 [SIMULATION] Extracting JWT token from Authorization Bearer header...")
    time.sleep(1)
    
    # Generate fresh token
    fresh_token = generate_fresh_token()
    print("✅ [SIMULATION] Fresh token extracted successfully!")
    
    # Step 5: Validate and save the new token
    print("\n📋 Step 5: Validate and save new token")
    
    # Validate the fresh token
    if token_manager._is_valid_token(fresh_token):
        print("✅ New token validation: PASSED")
        
        # Save to file
        if token_manager._save_token_to_file(fresh_token):
            print("💾 New token saved to file: SUCCESS")
            
            # Update internal state
            token_manager.current_token = fresh_token
            token_manager.last_refresh_time = datetime.now()
            
            print("🔄 Token refresh process: COMPLETED")
        else:
            print("❌ Failed to save token to file")
    else:
        print("❌ New token validation: FAILED")
        return False
    
    # Step 6: Verify the refresh worked
    print("\n📋 Step 6: Verify refresh success")
    new_info = token_manager.get_token_info()
    print(f"📊 New status: {new_info['status']}")
    print(f"⏰ New expires in: {new_info.get('expires_in', 'N/A')} seconds")
    print(f"📅 New expires at: {new_info.get('expires_at', 'N/A')}")
    
    if new_info['status'] == 'valid':
        print("🎉 TOKEN REFRESH SIMULATION: SUCCESS!")
        return True
    else:
        print("❌ TOKEN REFRESH SIMULATION: FAILED!")
        return False


def test_chatbot_collector_integration():
    """Test ChatbotVerbatimsCollector with refreshed token"""
    
    print("\n🤖 Testing ChatbotVerbatimsCollector Integration")
    print("=" * 60)
    
    # Test with auto-refresh enabled
    collector = ChatbotVerbatimsCollector(auto_refresh_tokens=True)
    
    # Test connection (this should trigger token validation)
    success, message = collector.test_connection()
    print(f"🔗 Connection test: {message}")
    
    if success:
        print("✅ ChatbotVerbatimsCollector integration: SUCCESS!")
        print("📝 The collector can now work with automatically refreshed tokens")
        
        # Show token info
        if hasattr(collector, 'token_manager') and collector.token_manager:
            token_info = collector.token_manager.get_token_info()
            print(f"📊 Final token status: {token_info['status']}")
            print(f"⏰ Token expires in: {token_info.get('expires_in', 'N/A')} seconds")
        
        return True
    else:
        print("❌ ChatbotVerbatimsCollector integration: FAILED!")
        return False


def main():
    """Run the complete simulation"""
    
    print("🧪 COMPLETE TOKEN REFRESH SIMULATION")
    print("=" * 80)
    print("This test simulates the complete automatic token refresh process")
    print("that would happen when Chrome automation extracts a fresh token.")
    print()
    
    try:
        # Run token refresh simulation
        refresh_success = simulate_token_refresh()
        
        if refresh_success:
            # Test integration
            integration_success = test_chatbot_collector_integration()
            
            if integration_success:
                print("\n🎉 SIMULATION COMPLETE: ALL TESTS PASSED!")
                print("💡 In a real Windows environment with Chrome:")
                print("   1. Chrome would open automatically when token expires")
                print("   2. The system would navigate to verbatims page")
                print("   3. Fresh JWT token would be extracted from network requests")
                print("   4. The token file would be updated automatically")
                print("   5. Your code would continue working seamlessly")
            else:
                print("\n❌ SIMULATION FAILED: Integration test failed")
        else:
            print("\n❌ SIMULATION FAILED: Token refresh failed")
            
    except Exception as e:
        print(f"\n💥 SIMULATION ERROR: {e}")
        import traceback
        traceback.print_exc()


if __name__ == "__main__":
    main()
