requirements.txt
# Core dependencies for dashboard analyzer
pandas>=1.5.0
numpy>=1.21.0

# JWT and HTTP requests for chatbot API
PyJWT>=2.8.0
requests>=2.28.0

# Selenium for browser automation and token extraction
selenium>=4.15.0
webdriver-manager>=4.0.1
undetected-chromedriver>=3.5.0

# Logging and utilities
python-dateutil>=2.8.0

# Optional: Power BI integration (if needed)
# powerbi-client>=1.0.0